chore: moved BigQuery helpers from artifacts pipeline to prepare for use in the review pipeline
