+++ pkg/bq/bigquery.go
@@ -12,14 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package artifact
+package bq
 
 import (
 	"context"
 	"errors"
 	"fmt"
-	"strings"
-	"text/template"
 
 	"cloud.google.com/go/bigquery"
 	"google.golang.org/api/iterator"
@@ -28,41 +26,9 @@ import (
 
 // BigQuery provides a client and dataset identifiers.
 type BigQuery struct {
-	projectID string
-	datasetID string
+	ProjectID string
+	DatasetID string
 	client    *bigquery.Client
-}
-
-// sourceQuery is the driving BigQuery query that selects events
-// that need to be processed.
-const sourceQuery = `
-SELECT
-	delivery_id,
-	JSON_VALUE(payload, "$.repository.full_name") repo_slug,
-	JSON_VALUE(payload, "$.repository.name") repo_name,
-	JSON_VALUE(payload, "$.repository.owner.login") org_name,
-	JSON_VALUE(payload, "$.workflow_run.logs_url") logs_url,
-	JSON_VALUE(payload, "$.workflow_run.actor.login") github_actor,
-	JSON_VALUE(payload, "$.workflow_run.html_url") workflow_url
-FROM {{.BT}}{{.ProjectID}}.{{.DatasetID}}.{{.EventTableID}}{{.BT}}
-WHERE
-event = "workflow_run"
-AND JSON_VALUE(payload, "$.workflow_run.status") = "completed"
-AND delivery_id NOT IN (
-SELECT
-  delivery_id
-FROM {{.BT}}{{.ProjectID}}.{{.DatasetID}}.{{.ArtifactTableID}}{{.BT}}
-)
-LIMIT {{.BatchSize}}
-`
-
-type queryParameters struct {
-	ProjectID       string
-	DatasetID       string
-	EventTableID    string
-	ArtifactTableID string
-	BatchSize       int
-	BT              string
 }
 
 // NewBigQuery creates a new instance of a BigQuery client.
@@ -73,8 +39,8 @@ func NewBigQuery(ctx context.Context, projectID, datasetID string, opts ...optio
 	}
 
 	return &BigQuery{
-		projectID: projectID,
-		datasetID: datasetID,
+		ProjectID: projectID,
+		DatasetID: datasetID,
 		client:    client,
 	}, nil
 }
@@ -90,26 +56,8 @@ func (bq *BigQuery) Close() error {
 // Query takes a queryString (assumed to be valid SQL) and executes it against
 // BigQuery using the given client. The results are then mapped to a slice of T,
 // where each row in the result is mapped to a struct of type T.
-func Query[T any](ctx context.Context, bq *BigQuery, eventsTable, artifactTable string, batchSize int) ([]*T, error) {
-	tmpl, err := template.New("query").Parse(sourceQuery)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse query template: %w", err)
-	}
-
-	var sb strings.Builder
-	if err := tmpl.Execute(&sb, &queryParameters{
-		ProjectID:       bq.projectID,
-		DatasetID:       bq.datasetID,
-		EventTableID:    eventsTable,
-		ArtifactTableID: artifactTable,
-		BatchSize:       batchSize,
-		BT:              "`",
-	}); err != nil {
-		return nil, fmt.Errorf("failed to apply query template parameters: %w", err)
-	}
-	fmt.Println(sb.String())
-
-	query := bq.client.Query(sb.String())
+func Query[T any](ctx context.Context, bq *BigQuery, queryString string) ([]*T, error) {
+	query := bq.client.Query(queryString)
 	job, err := query.Run(ctx)
 	if err != nil {
 		return nil, fmt.Errorf("query.Run failed: %w", err)
@@ -133,8 +81,8 @@ func Query[T any](ctx context.Context, bq *BigQuery, eventsTable, artifactTable 
 	return items, nil
 }
 
-func Write(ctx context.Context, bq *BigQuery, tableID string, artifacts []*ArtifactRecord) error {
-	if err := bq.client.Dataset(bq.datasetID).Table(tableID).Inserter().Put(ctx, artifacts); err != nil {
+func Write[T any](ctx context.Context, bq *BigQuery, tableID string, rows []*T) error {
+	if err := bq.client.Dataset(bq.DatasetID).Table(tableID).Inserter().Put(ctx, rows); err != nil {
 		return fmt.Errorf("failed to write to BigQuery: %w", err)
 	}
 	return nil
